Add a log line to see how often the S3 cache is available
AWS Lambda has _some_ global state, and we are using it to increase performance.
This state is tied to the execution environment, and is therefore a little tricky to know how reliable it is.

In this change we add a log line to help us determine how often cached data is being used.