feat!: Remove ability to assume a role to write to kinesis
Remove the optional `KinesisStreamRole` parameter from the CloudFormation template.

An analysis (see below) of instances of this stack show the following accounts have this set:
  - workflow
  - media-service
  - cmsFronts

These stacks belong to the Editorial Tools team.

Only they have set the parameter because the logging setup within Editorial Tools was a little different from most.
The team used to have a single logging stream that was shared between their various AWS accounts;
apps that ran in a different account from the stream assumed a role to gain permission to write to it.

The team have since updated this, favouring a stream per account.
This can be confirmed by observing the value of the `KinesisStreamArn` parameter - the value maps to a stream in the account.

Analysis performed via:

```sh
\#!/usr/bin/env bash

set -e

REPOSITORY_NAME=guardian/cloudwatch-logs-management

\# source from riff-raff.yaml and match it to Janus' naming of accounts
PROFILES=(
  composer
  workflow
  media-service
  capi
  cmsFronts
  ophan
  frontend
  identity
  mobile
  deployTools
  targeting
  investigations
)

for profile in "${PROFILES[@]}"; do
  aws cloudformation describe-stacks \
    --profile "$profile" \
    --region eu-west-1 \
    --no-paginate | \
    jq ".Stacks[] | select(.Tags[].Value == \"${REPOSITORY_NAME}\") | .Parameters"
done
```