refactor: Add explicit handlers directory for clarity
In this change, we create a `handlers` directory to house the entry points for the lambdas,
in the hope of increasing maintainability after observing it being difficult to know where to start.

It's worth noting that the layout of this service is different from most others.
This service produces a mono-zip with multiple entry points.
An alternative to this change could be to produce an artifact per lambda.