Move config fetching into lambda handlers
Before config was fetched at the top level of each of the files which
define the lambda handlers. This means that the config was fetched when
the file was included, even if the handler wasn't run. So when running
the set-retention lambda, for example, we were seeing errors because the
config for the ship-log-entries lambda wasn't available.