Update README.md to clarify when we mean deployment-stack rather than cloudformation stack.
The word "stack" is horribly overloaded and most folk I speak to outside of devX see it exclusively in the context of Cloudformation Stacks. This can lead to a lot of confusion.
The readme was using "stack" in both the cloudformation sense and also in the riffraff deployment-resource sense. I am hoping this rewording and the addition of a link to the riffraff deployment-resource-stacks list will make this a bit clearer.